Fix segfault when using libxml2
Should not do a general, global memory cleanup in a local delete function.
The libxml2 library has a global list of schema type pointers that are
freed when calling xmlCleanupParser().  If other code, such as opensync
itself, is still relying on these schema types, a segfault will occur.

Leave the global cleanups for the applications.

git-svn-id: http://svn.opensync.org/plugins/google-calendar@6101 53f5c7ee-bee3-0310-bbc5-ea0e15fffd5e