Reworked the state_db flags, storing both edit_url and etag for modification
During testing, it appears that it is necessary to supply to Google
both the edit_url and the etag in order to modify an entry, and just
the edit_url to delete it.

Also of interest, when adding or modifying an existing entry, the response
from Google does not include the original ID, but puts the edit URL in
its place.

So it is not possible to rely on the ID always being there, even though
Google documentation asks for it to be there.  Fortunately, it seems to
work with just edit URL and etag.

This large patch updates both modify and delete logic, as well as supporting
code, to accommodate this requirement, using the edit URL as the state_db
key (again), and the etag as the data.

git-svn-id: http://svn.opensync.org/plugins/google-calendar@6111 53f5c7ee-bee3-0310-bbc5-ea0e15fffd5e